requirements.txt
aiogram==3.23.0

python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0

asyncpg==0.31.0
sqlalchemy[asyncio]==2.0.45
alembic==1.17.2
psycopg[binary]==3.3.2

APScheduler==3.11.2

python-dateutil==2.9.0.post0
pytz==2025.2
